refactor drawROIWindow
